codegen: relocate Solana sources into targets/solana/
