+++ src/codegen/dispatch/mod.rs
@@ -1,4 +1,3 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub(crate) mod polkadot;
-pub(super) mod solana;
