+++ src/codegen/targets/solana/events.rs
